sink list page: ask to unload module-policy-enforcement
module-policy-enforcement overrides the default sink set by the
PulseAudio server, basically making that setting ineffective. The simple
workaround for now is to offer unloading the policy enforcement module
when a default sink is set. In the long run though, we'll need to do
something more sensible, e.g. utilizing roles.